+++ projects/deja-js/component/CHANGELOG.md
@@ -1,3 +1,26 @@
+# [14.4.0-develop.4](https://github.com/DSI-HUG/dejajs-components/compare/v14.4.0-develop.3...v14.4.0-develop.4) (2023-06-09)
+
+# [14.4.0-develop.3](https://github.com/DSI-HUG/dejajs-components/compare/v14.4.0-develop.2...v14.4.0-develop.3) (2023-05-12)
+
+
+### Bug Fixes
+
+* fix range binding ([7317f2c](https://github.com/DSI-HUG/dejajs-components/commit/7317f2cab9c0a320e66c2b34f8ce036b2915476d))
+
+# [14.4.0-develop.2](https://github.com/DSI-HUG/dejajs-components/compare/v14.4.0-develop.1...v14.4.0-develop.2) (2023-05-12)
+
+
+### Features
+
+* range, add trackby for range component ([e1bcf5f](https://github.com/DSI-HUG/dejajs-components/commit/e1bcf5fa7ee67e5d8aa15d9cb7089f226b8f5d1b))
+
+# [14.4.0-develop.1](https://github.com/DSI-HUG/dejajs-components/compare/v14.3.0...v14.4.0-develop.1) (2023-05-12)
+
+
+### Features
+
+* tiles.component.html, add tryBy property ([72b495e](https://github.com/DSI-HUG/dejajs-components/commit/72b495eca745520455a5a55896480078838bd4a0))
+
 # [14.3.0](https://github.com/DSI-HUG/dejajs-components/compare/v14.2.0...v14.3.0) (2023-04-20)